check-s3.sh

+++ check-s3.sh
@@ -9,12 +9,15 @@ do
 	echo "Working on $bucket ..."
     SIZE="$(aws s3 ls s3://$bucket --recursive --summarize 2> /dev/null | grep Size | awk '{print $3}')"
 
-    if [ $SIZE -gt 20971520 ]
+    if [ $SIZE != "" ]
     then
-        echo "Emptying $bucket ..."
-        aws s3 rm s3://$bucket --recursive
-    else
-        echo "Bucket size under 20 MiB ($SIZE Bytes) ..."
+        if [ $SIZE -gt 20971520 ]
+        then
+            echo "Emptying $bucket ..."
+            aws s3 rm s3://$bucket --recursive
+        else
+            echo "Bucket size under 20 MiB ($SIZE Bytes) ..."
+        fi
     fi
 done < "${in}"
